# backend-apirest/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Presencia Backend API REST UAT
  version: 0.1.0
  description: >
    API REST interna para autenticar contra el portal escolar UAT con cookies
    ASP.NET y consultar horarios/examenes de profesor.
servers:
  - url: http://localhost:3100
security: []
paths:
...
  /api/uat/sessions:
    post:
      summary: Crea una sesion autenticada contra UAT
      description: >
        Ejecuta GET /Login, POST /Login/Accesar_Dominio y GET /Login/Validar
        manteniendo un CookieJar persistente en memoria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Sesion UAT creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
  /api/uat/alumnos/sessions:
    post:
      summary: Crea una sesion autenticada contra el portal de alumnos SUR
      description: >
        Obtiene __RequestVerificationToken, ejecuta Login/Accesar_Dominio,
        valida la sesion, consulta carreras y selecciona el plan solicitado o
        el primero disponible.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentCredentials"
      responses:
        "201":
          description: Sesion de alumno creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSessionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    CoordinatorCookie:
      type: apiKey
      in: cookie
      name: coord_session
...
  responses:
    Error:
      description: Error de validacion, autenticacion o portal
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Credentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: profesor@uat.edu.mx
        password:
          type: string
          format: password
    StudentCredentials:
      allOf:
        - $ref: "#/components/schemas/Credentials"
        - type: object
          properties:
            idPlanEstudio:
              type: integer
              minimum: 1
              example: 3313
            attendanceUuid:
              type: string
              format: uuid
              description: UUID BLE estable generado por la app de alumnos.
            deviceBindingId:
              type: string
              format: uuid
              description: Identificador estable de la instalacion/dispositivo.
            platform:
              type: string
              example: android
            deviceInfo:
              type: string
...
    SessionResponse:
      type: object
      required: [sessionId, authenticated, login, createdAt, lastUsedAt, expiresAt]
      properties:
        sessionId:
          type: string
          format: uuid
        authenticated:
          type: boolean
        login:
          $ref: "#/components/schemas/SafeLogin"
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        activeSessions:
          type: integer
        cookieDiagnostics:
          type: object
          properties:
            cookieNames:
              type: array
              items:
                type: string
            hasSessionCookie:
              type: boolean
            hasAuthCookie:
              type: boolean
    StudentSessionResponse:
      allOf:
        - $ref: "#/components/schemas/SessionResponse"
        - type: object
          required: [careers, selectedCareer]
          properties:
            careers:
              type: array
              items:
                type: object
                additionalProperties: true
            selectedCareer:
              type: object
              additionalProperties: true
    SafeLogin:
      type: object
      properties:
        exito:
          type: boolean
        cambiaPass:
          type: boolean
        mensaje:
          type: string
        parametros:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: UAT_LOGIN_FAILED
        message:
          type: string
        details:
          nullable: true